gae-product-service/requirements.txt
apache-beam==2.2.0
avro==1.8.2
beautifulsoup4==4.6.0
cachetools==2.0.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
crcmod==1.7
dill==0.2.6
enum34==1.1.6
Flask==0.12.2
funcsigs==1.0.2
future==0.16.0
futures==3.2.0
gapic-google-cloud-pubsub-v1==0.15.4
google-api-python-client==1.6.5
google-apitools==0.5.11
google-auth==1.4.1
google-auth-httplib2==0.0.3
google-cloud-bigquery==0.25.0
google-cloud-core==0.25.0
google-cloud-dataflow==2.2.0
google-cloud-pubsub==0.26.0
google-gax==0.15.16
googleapis-common-protos==1.5.3
googledatastore==7.0.1
grpc-google-iam-v1==0.11.4
grpcio==1.9.1
httplib2==0.9.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mock==2.0.0
oauth2client==3.0.0
pbr==3.1.1
ply==3.8
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-pubsub-v1==0.15.4
protobuf==3.3.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
six==1.10.0
typing==3.6.4
uritemplate==3.0.0
urllib3==1.22
Werkzeug==0.14.1
